minor: refactor stores/queries code
Improve a bit the code as it was quite hard to follow
